+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support for `for...in` loops and `available` checks to result builder APIs.
+
 ### Removed
 
 ### Changed